Adding new Component to Sample Changer ui
