+++ bsx3/ui/app/redux-store.js
@@ -1,5 +1,3 @@
-/* eslint-disable indent */
-/* eslint-disable eol-last */
 import { createStore, applyMiddleware, compose } from 'redux';
 import { routerMiddleware } from 'connected-react-router';
 import logger from 'redux-logger';
@@ -14,29 +12,29 @@ const enhancers = [];
 
 // Logger MUST BE the last middleware
 const middleware = [
-    thunk,
-    routerMiddleware(history),
-    logger
+  thunk,
+  routerMiddleware(history),
+  logger
 ];
 
 if (process.env.NODE_ENV === 'development') {
-    // eslint-disable-next-line no-underscore-dangle
-    const devToolsExtension = window.__REDUX_DEVTOOLS_EXTENSION__;
+  // eslint-disable-next-line no-underscore-dangle
+  const devToolsExtension = window.__REDUX_DEVTOOLS_EXTENSION__;
 
-    if (typeof devToolsExtension === 'function') {
-        enhancers.push(devToolsExtension());
-    }
+  if (typeof devToolsExtension === 'function') {
+    enhancers.push(devToolsExtension());
+  }
 }
 
 const composedEnhancers = compose(
-    applyMiddleware(...middleware),
-    ...enhancers
+  applyMiddleware(...middleware),
+  ...enhancers
 );
 
 const store = createStore(
-    createRootReducer(history),
-    initialState,
-    composedEnhancers
+  createRootReducer(history),
+  initialState,
+  composedEnhancers
 );
 
 export default store;
