+++ bsx3/ui/app/root-reducer.js
@@ -3,7 +3,8 @@ import { connectRouter } from 'connected-react-router';
 
 import app from './actions/app';
 import beamline from './actions/beamline';
-import sampleChanger from './actions/sampleChanger';
+import sample from './actions/scSample';
+import buffer from './actions/scBuffer';
 import queue from '../views/datacollection/queue/queue-api';
 import login from './actions/login';
 
@@ -12,7 +13,8 @@ export default function createRootReducer(history) {
     router: connectRouter(history),
     app,
     beamline,
-    sampleChanger,
+    sample,
+    buffer,
     queue,
     login,
   });
