+++ bsx3/ui/components/propertyButton/PropertyButton.jsx
@@ -9,15 +9,6 @@ class PropertyButton extends React.Component {
     // let notif = false;
     return (
       <div className="flexprop">
-        {/* {notif
-          ? (
-            <Confirmation
-              show
-              message=" Are you sure you want to delete this Record?"
-            />
-          )
-          : null
-      } */}
         <Button
           variant="link"
           title="Edit row"
@@ -32,7 +23,6 @@ class PropertyButton extends React.Component {
           title="Delete row"
           className="edit-del del-btn"
           onClick={() => {
-            // notif = true;
             if (window.confirm(`Are you sure you want to delete this row? ${this.props.row.id} ${this.props.showNotification} `)) {
               this.props.handleDeleteRow(this.props.row);
               this.props.handleShowNotification(true);
