+++ bsx3/ui/components/table/TableRow.jsx
@@ -19,7 +19,7 @@ export default class TableRow extends Component {
       >
         <Item>Add to Queue</Item>
         <Separator />
-        <Item onClick={() => this.props.handleDuplicateRow(this.props.row.id)}>
+        <Item onClick={() => this.props.handleDuplicateRow(this.props.row)}>
           <span>
             Duplicate
             { ' ' }
@@ -33,7 +33,7 @@ export default class TableRow extends Component {
             <i className="far fa-edit" />
           </span>
         </Item>
-        <Item onClick={() => this.props.handleDeleteRow(this.props.row.id)}>
+        <Item onClick={() => this.props.handleDeleteRow(this.props.row)}>
           <span>
             Delete row
             { ' ' }
