+++ bsx3/ui/views/datacollection/sampleChanger/style.css
@@ -7,20 +7,33 @@
 .wrapsc{
     margin-top: 40px;
     margin-left: 10px;
-    /* border: 1px solid black;  */
 }
+.optimiz{
+    display: flex;
+    background: rgba(93, 146, 56, 0.164);
+    border: 1px solid lightgray; 
+    border-radius: 5px; 
+    height: 40px;
+    margin-right: 15px
 
+}
 .divcollapsebtn{
  width: 193px;
- background: rgba(108, 145, 81, 0.164);
+ background: rgba(93, 146, 56, 0.164);
  border: 1px solid lightgray; 
- border-radius: 20px; 
+ border-radius: 5px; 
 }
 
 .collapsebtn{
+    -moz-appearance: none;
     border: 2px solid black; 
-    border-radius: 20px!important; 
+    /* color: rgb(124, 120, 120)!important;  */
 }
+
+.collapsebtn:focus{
+    box-shadow: none!important; 
+}
+
 
 .panel-heading{
     background: rgba(108, 145, 81, 0.144);
@@ -82,11 +95,6 @@
     background-color: #8bc34a;
   }
 
-  
-  input:focus + .slider {
-    box-shadow: 0 0 1px #2196F3;
-  }
-  
   input:checked + .slider:before {
     -webkit-transform: translateX(13px);
     -ms-transform: translateX(13px);
